sc_copilot/requirements.txt
numpy
google-generativeai
configparser
